# Refactor struct Converter and reduce template parameters