--- a/velox/functions/prestosql/ArrayFunctions.h
+++ b/velox/functions/prestosql/ArrayFunctions.h
@@ -206,7 +206,8 @@ struct ArrayJoinFunction {
   template <typename C>
   void writeValue(out_type<velox::Varchar>& result, const C& value) {
     result +=
-        util::Converter<TypeKind::VARCHAR, void, false, false>::cast(value);
+        util::Converter<TypeKind::VARCHAR, void, util::DefaultCastPolicy>::cast(
+            value);
   }
 
   template <typename C>