--- a/velox/type/tests/ConversionsTest.cpp
+++ b/velox/type/tests/ConversionsTest.cpp
@@ -42,13 +42,14 @@ class ConversionsTest : public testing::Test {
 
     auto cast = [&](TFrom input) -> TTo {
       if (truncate & legacyCast) {
-        return Converter<toTypeKind, void, true, true>::cast(input);
+        return Converter<toTypeKind, void, TruncateLegacyCastPolicy>::cast(
+            input);
       } else if (!truncate & legacyCast) {
-        return Converter<toTypeKind, void, false, true>::cast(input);
+        return Converter<toTypeKind, void, LegacyCastPolicy>::cast(input);
       } else if (truncate & !legacyCast) {
-        return Converter<toTypeKind, void, true, false>::cast(input);
+        return Converter<toTypeKind, void, TruncateCastPolicy>::cast(input);
       } else {
-        return Converter<toTypeKind, void, false, false>::cast(input);
+        return Converter<toTypeKind, void, DefaultCastPolicy>::cast(input);
       }
     };