Refactor struct Converter in Conversions.h to remove the code
complexity from the template parameter TRUNCATE and LEGACY_CAST.
Wrap them into a single template parameter TPolicy as well as
any future additional parameter.

struct Converter holds the core logic of the CAST expression
between scalar/primitive types. Benchmarking shows ignorable
impact on perf.

Also refactor the case of ToKind being BOOLEAN, and move the logic
into a single template specialization. The logic used to spread
over two template specializations, for TRUNCATE true and false.