@gggrace14 2023-12-07
Benchmarking results -- No significant change to CAST perf

**Before**
> ============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          1.67ms    599.75
cast##tryexpr_cast_invalid_empty_input                    334.93ms      2.99
cast##try_cast_invalid_nan                                359.54ms      2.78
cast##tryexpr_cast_invalid_nan                            740.74ms      1.35
cast##try_cast_valid                                        2.05ms    488.87
cast##tryexpr_cast_valid                                    2.05ms    487.40
cast##cast_valid                                            2.05ms    488.59
cast##cast_decimal_to_inline_string                         7.38ms    135.57
cast##cast_short_decimal                                   11.70ms     85.47
cast##cast_long_decimal                                    27.60ms     36.24
cast##cast_large_real_to_scientific_notation               22.12ms     45.22
cast##cast_small_real_to_standard_notation                  9.08ms    110.16
cast##cast_small_double_to_scientific_notation             21.36ms     46.83
cast##cast_large_double_to_standard_notation               11.29ms     88.60
cast##cast_timestamp                                       11.10ms     90.11


**After**
> ============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          1.71ms    584.94
cast##tryexpr_cast_invalid_empty_input                    324.43ms      3.08
cast##try_cast_invalid_nan                                351.58ms      2.84
cast##tryexpr_cast_invalid_nan                            719.29ms      1.39
cast##try_cast_valid                                        2.06ms    485.46
cast##tryexpr_cast_valid                                    2.07ms    482.95
cast##cast_valid                                            2.03ms    493.19
cast##cast_decimal_to_inline_string                         7.04ms    142.08
cast##cast_short_decimal                                   11.34ms     88.20
cast##cast_long_decimal                                    26.88ms     37.21
cast##cast_large_real_to_scientific_notation               21.13ms     47.32
cast##cast_small_real_to_standard_notation                  8.86ms    112.93
cast##cast_small_double_to_scientific_notation             20.72ms     48.26
cast##cast_large_double_to_standard_notation               10.88ms     91.93
cast##cast_timestamp                                       11.22ms     89.10